requirements.txt
# Microsoft Graph API
msal>=1.24.0
requests>=2.31.0

# Excel handling
openpyxl>=3.1.2
pandas>=2.0.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
pdfplumber>=0.10.0

# Date/Time
python-dateutil>=2.8.2
pytz>=2023.3

# Configuration
python-dotenv>=1.0.0

# Claude API (optional)
anthropic>=0.18.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.1

# Type hints
typing-extensions>=4.7.0
